ENH: Add mappings to paste with one number (which depending on the current cursor position) incremented / decremented
Handle repeat of gpp with the last used offset and the same number position by introducing a special ".p" paste type.
FIX: Don't lose the original [count] given when repeating the mapping. As UnconditionalPaste#Paste() executes a normal mode command, we need to store v:count and make it available to the <Plug>-mapping via the new UnconditionalPaste#GetCount() getter.