Don't add Oxford comma when exactly two lines
Either if there are two lines in the register, or a single line with a count of 2. This should result in "Tom and Jerry" instead of "Tom, and Jerry". I think it's far more useful to be able to paste two lines with a conjunction in grammatical correct form, than supporting the use case of appending the last two elements to an existing comma-separated list.