ENH: Handle separators around paste in insert and command-line modes, too
(So far, we'd supported separator pastes only via the existing normal mode mappings; now, with the universal Combinatorial mapping, they may be used in the other modes, too.)
Need to get the current mode into the algorithm, so add a:mode to UnconditionalPaste#Insert() and UnconditionalPaste#Paste(), then forwarded to s:ApplyAlgorithm().
Make UnconditionalPaste#Separators#Check() a switch on the mode; for insert mode, we can mostly reuse the implementation (now in s:BufferCheck()), but with a:isPasteAfter always set to 0, as the cursor is on the following character, so from a positional perspective, we're always pasting before the cursor position (even though in insert mode, the text will always be appended). For the determining of whether prefix / suffix shall be used, leave off the a:isPasteAfter check altogether (like for command-line mode).
For command-line mode, the getline() and search()es won't work, we have to go with getcmdpos() and getcmdline() instead. The whole a:isPasteAfter doesn't apply here.