ENH: Handle indent and linewise in insert mode, too
We assume that the indent settings in insert mode work and therefore only indent / dedent the first line (by prepending <Tab> / <BS> to the returned pasteContent (which with CTRL-R are interpreted as typed); the rest is taken care of by indenting.
For linewise pastes, the trailing newline at the end already creates the trailing empty line, but at the beginning, we need another newline to create an empty line.