Make insert mode Indent (>) always start from column 1
Otherwise, the insert mode behavior would be identical to MoreIndent (m). We let the algorithm prepend a <C-U> so that the cursor (in linewise paste) returns to the start of the line, clearing any auto-indent. This interferes with the other prependings in UnconditionalPaste#Paste(), so we have to temporarily split off the <C-U> and rejoin it later at the right position (after the initial <CR> that starts a new line, but before any additional shifts).