Reuse ingo#subs#apply#FlexibleExpression() for the g=p mapping
To support application of the expression not just to lines, but the entire text, or cutsom matches or splits. With this, function names, external and Ex commands can also be used, making the "expression" truly flexible.
By reusing the functionality from the ingo-library, feature parity and compatibility to the very related g= mapping from SubstituteExpression.vim is obtained.
Adapt the tests to maintain the original intention. Only minimal changes are applied to reflect the changed behavior.