Join g=p returned List with newlines
When the queried expression is applied to the entire register contents, a returned List reaches the plugin. Behave like SubstituteExpression.vim in that case: Treat List elements as individual lines and join with newlines. If elements themselves are already linewise (so have a trailing newline), don't add an empty line, though.